Fix for #12945: ProgressBar.setMax() will always redraw.
Fix for issue #12945: Changing the maximum of a progress bar does
not cause it to be redrawn, even though a new maximum changes the
position of the current progress in relation to its maximum. With this fix,
setMax() will always cause refreshProgress() to be called if the maximum is
different than it was before.

Change-Id: Ice2c28916dc609124b0f3e4c981a3e83d2245361