TextView: don't show context menu for double-tap
Use the mechanism long-press uses to prevent the context menu
from being shown. hide()ing the selection was not working properly.

Also, prevent a triple-tap from being seen as two double-tap events.

BEWARE: Google has apparently already fixed this internally in 2.3.4,
but we do not have the source for that version at this time, nor is it
clear that the 2.3.4 fix will be available in time for a final releas

Change-Id: Ic99c494b285003a5aa8610e921aeaced4cbdc9b0