Fw: Theme compatibility
 * Added in database entries for hasModdedBattery & hasModdedSignal.
to be used in the AndroidManifest.xml for compatible themes as follows.

        pluto:hasModdedBattery="true"
        pluto:hasModdedSignal="true"

This will set each value to true and then when each theme is applied if
those have been set to true

THEME_COMPATIBILITY_BATTERY &
THEME_COMPATIBILITY_SIGNAL

will be set to true, allowing for 6bar and 100 battery mods. If they are
not set in the AndroidManifest, they will default to false.

I also changed the mThemeCompatibility variables to booleans instead of
integers. Made more sense to me.

Change-Id: I83cfdf797545da871e0ac85b7dc9bf63525c364f