Fix for PduParser.java, allows mms recieve on supersonic & speedy
mms recieve on supersonic is not working due to changes
in content disposition parsing in PduParser.java.
This change fixes the if statement which previously ran
if the preference config_mms_content_disposition_support
was true, which it always was.
This change also needs the preference in the device
specific overlay config.xml to be set to false for
affected devices/carriers (Sprint).

Change-Id: Ic8c074fe19daf3832ad65cef64c50a4299ac2496