Pass the path to format to ExternalStorageFormatter.
We need to pass the path to the device we want to
format down to ExternalStorageFormatter. It should
not use the default of getExternalStoragePath()
because some devices have both an sdcard and an emmc.
If it is not passed down, the device always
formats the sdcard even if it is emmc that needs
formatting.

Change-Id: I8185a63eaefc35d15ed487c6c4100252f0441524