libstagefright: Crop the native window with the display resolution
We were providing the native window with the display width and height in
set_geometry.  On the surfaceflinger side, they assume that this is
aligned, which causes problems for HDMI.  Instead we'll now pass in the
aligned stride and slice in native_window_set_geometry, then specify the desired
display width and height by calling native_window_set_crop.

Change-Id: I9b9a910e951468bdeb30f3f8ea9df70f3c2ad00a
CRs-Fixed: 283063

Conflicts:

	media/libstagefright/OMXCodec.cpp

Signed-off-by: Andrew Sutherland <dr3wsuth3rland@gmail.com>