added ability to keep Parcel::appendFrom GB compat
in ICS, the first argument of Parcel::appendFrom was changed to const.
This causes linker problems when blobs from GB ROMs try to link with
libbinder.so.

to enable, set BOARD_KEEP_PARCEL_GB_COMPATIBLE:=true in BoardConfig.mk

this is specifically needed for Skyrocket, but may be useful for other
ICS ports with GB blobs

Change-Id: I22c412706ab2ab252f4897eae171768d73ec2bf8