MPEG4Writer: add workaround for passion
For an unknown reason passion (or at least my passion) is mangling
previousPausedDuration on one of the loops creating a negative timestampUs
which causes the CHECK to fail. So if we expect the CHECK to fail replace
previousPausedDurationUs with a sane value. In testing this seems to always
work, creating playable/seekable videos with proper timestamp info.

Change-Id: I4795cd000b80315b6388875eb8dacff01dd684d9