NotificationManagerService: Fix notification led bug (Part 1/2)
We can assume that if the user turned the screen off while there was
still an active notification then they wanted to keep the notification
for later. In this case we shouldn't flash the notification light.

The problem for some special notifications (such as missed calls)
is that the screen is automatically turned on, and then when it is
turned back off the led does not get enabled. For these notifications
we use a flag to force the notification light even if the screen was
turned off.

Change-Id: Ib8185df93ef822fcf9386ca504b05ad4da1d1295

Conflicts:
	services/java/com/android/server/NotificationManagerService.java