NLP Combo feature
If Combo Provider is listening for network
locations, it will want to screen these locations
before they are sent out to clients.

Get the package name for combo provider from resouce
instead of hardcoding the name. Also, geofence services
provider is renamed to differentiate from the hal
geofence provider

Only screen network locations if Combo provider is
the connected network location provider.

Resolved issue where MockProvider was being cast to
LocationProviderProxy and would cause a fatal exception

CRs-fixed: 483398, 608477, 619882

Change-Id: I3156e85307b40aee9fb06b433c83153c84b8ed42

Compilation fix for mPackageName

Change-Id: I9444a73d1b2c0c9a098f08710e26af62032b4c76

combo bug when location reported by providers do not come with extras

LocationManagerService tries to create an extra bundle and stick it
into the location object, but it is not doing it correctly. As a
result, combo drops report.

Change-Id: I6f5886dd7f14819ecc7a4ed96c867a0ec486faf7
CRs-Fixed: 962746