Add upgrade step to fix system sounds issue
With Q release, default sounds were moved to /product.
However, the settings update was not provided, and as a
result, the upgrade P -> Q  breaks all the default system
sounds that are kept in memory and persisted to XML file.

Apparently, there are no way to handle this other
than resetting settings of all of these sounds...

Test: After upgrade lock/unlock sounds are present,
charging sound is correct.

Change-Id: I43c2cbf53bca856b4659ab18c6d6e4cc77ab0413