core: spoof device information for certain applications
 * Spoof build fingerprint for Google Play Services

SafetyNet's CTS profile attestation checks whether Build.FINGERPRINT
matches that of the device's stock OS, which has passed CTS testing.
Spoof the fingerprint for Google Play Services to help pass SafetyNet.

We used to set the real system build fingerprint to the stock one, but
Android relies on each build having a unique fingerprint in order to
clear the correct caches and update persistent state for system changes.
On devices that no longer receive updates from the OEM, the build
fingerprint never changes and Android doesn't account for updates
correctly, which causes issues when updating without wiping data.
Only spoofing the fingerprint for Google Play Services fixes this issue.

 * Defy usage of NGA in Google Assistant

Next-Generation Assistant (NGA) is nice to have for Google Assistant,
but it requires proprietary code in SystemUI to work properly. However,
newer Pixel devices (Pixel 4 and 5 series) automatically use NGA with no
way to opt out, which breaks the assistant entirely on custom ROMs
without NGA support.

Report the device as a "Pixel 3 XL" to the Google Assistant app during
app process specialization to force the usage of the generic assistant
UI and mitigate the issue. This is far from ideal, but we just need to
make Google Assistant work for now.

Screen-off voice match is not broken with this change because the model
name still matches that of a valid device with the feature.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Change-Id: I415b6760ecf4032bd60f886c319d43bc6d63aef6

Zygote: Override fingerprint for SettingsIntelligence

Change-Id: Ia839345da18dde46bfd1acb77b42e0c47d28c8ac

core: Make build property spoofing more reliable

Determining the package name from Zygote's post-fork process name is
unreliable for several reasons:
  - If USAP is enabled, Zygote will fork with the USAP process name, and
    spoofing will never trigger when the forked process specializes into
    a specific app.
  - With the native fork loop backported from AOSP master,
    Zygote.setAppProcessName() is no longer used for most apps because
    the forking is handled by native code.

Instrumentation#newApplication() is responsible for creating the
Application singleton during each app's initialization, so hook it
instead for a more reliable and robust trigger.

Change-Id: Ic762d8853967bf78db614efaf46166e527cb397a